refactor: change file names to kebab-case
