+++ class-components/src/main.tsx
@@ -1,8 +1,8 @@
 import { StrictMode } from 'react';
 import { createRoot } from 'react-dom/client';
 
-import { App } from './app/App.tsx';
-import { ErrorBoundary } from './utils/ErrorBoundary.tsx';
+import { App } from './app/app.tsx';
+import { ErrorBoundary } from './utils/error-boundary.tsx';
 
 import './global.css';
